Updates to User Service
